Various changes to prevent narrowing conversions: media/
Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2958308
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892143}

+++ media/mojo/common/mojo_shared_buffer_video_frame_unittest.cc
@@ -262,10 +262,10 @@ TEST(MojoSharedBufferVideoFrameTest, InterleavedData) {
   gfx::Rect visible_rect(size);
 
   // Create interlaced UV data, which are each 1/4 the size of the Y data.
-  const size_t y_offset = 0;
-  const size_t u_offset =
+  const uint32_t y_offset = 0;
+  const uint32_t u_offset =
       VideoFrame::PlaneSize(format, VideoFrame::kYPlane, size).GetArea();
-  const size_t v_offset =
+  const uint32_t v_offset =
       u_offset + VideoFrame::RowBytes(VideoFrame::kUPlane, format, kWidth);
   const int32_t y_stride =
       VideoFrame::RowBytes(VideoFrame::kYPlane, format, kWidth);
